requirements.txt
flask
numpy
opencv-python
mediapipe
Pillow
onnxruntime
scipy
open3d
trimesh
glb
pytorch3d
PyFBX
